Better handling of IRC PART events
The "Leave channel" menu items now actually issue an IRC PART command,
and we leave the subsequent UI changes up to the ChannelView and
ChannelButtonView. This means "/part" commands sent via the text input
work just like the GUI controls.

This commit removes the ability to remove a channel from the sidebar
(and remove the ChannelView itself from the DOM). Hopefully we can
reintroduce that later.