Abstract various menus into multi-purpose NativeMenu class
The new module NativeMenu can be used to create an application menu
(at the top of the screen on a Mac) or a context menu (like those
shown on right-click). It has a .construct() method that allows you
to completely replace its contents at any time, which means the
same NativeMenu instance can be re-used for a number of purposes,
with the contents being replaced on-demand to suit the situation.

NativeMenu is now used for the main application menu, and the
channelButtonView right-click menu. It will also, soon, be used
in the spelling suggestion right-click menu (#46) and the standard
text selection right-click menu (#47).