Monkeypatch new .joinChannel() method onto node-irc client instances
A more sensible solution to the lack of channel password support
in node-irc. .joinChannel() takes an optional password parameter,
or uses a saved password if not supplied. It also saves all
passwords it sees, so even channels joined manually (with a `/join`
text message) can be left and rejoined without a password from
then on.

Part of #53.